requirements.txt
openai>=1.44.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
rapidfuzz>=3.9.1
python-dotenv>=1.0.1
rich>=13.7.1
pydantic>=2.8.2
tabulate>=0.9.0

